Time Card Sync Feature
Add a new feature of syncing out time entries to Touchpoint's system on daily basis.

The "TimeCardSyncScheduler" will run every day to call the TimeCardSyncService to export/import our time entries to Touchpoint's system.

The "TimeCardSyncSetting" is a custom settings object, which includes the basic settings for the integration, such as the TeamMember Name, Email and Endpoint.

A custom field called "ExternalID" is added to Change object.

A custom field called "IsSync" is added to TimeCard object.